+++ packages/core/src/types/index.ts
@@ -166,7 +166,7 @@ export type {
   IViewport,
   ViewReference,
   ReferenceCompatibleOptions,
-  ViewReferenceSpecifier as ViewReferenceSpecifier,
+  ViewReferenceSpecifier,
   StackViewportProperties,
   VolumeViewportProperties,
   ViewportProperties,